requirements.txt
Flask>=0.12.2
Markdown>=3.3.4
python-dotenv>=1.0.1
Requests>=2.32.3
